# docs/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: TalentTrust API
  description: TalentTrust API Gateway and services documentation
servers:
  - url: /api/v1
paths:
  /contracts:
    get:
      summary: List contracts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
        "401":
          description: Unauthorized — missing or invalid JWT
        "403":
          description: Forbidden — insufficient permissions
    post:
      summary: Create a contract
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContract"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "400":
          description: Validation error
        "401":
          description: Unauthorized — missing or invalid JWT
        "403":
          description: Forbidden — role does not have contracts:create permission
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Contract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        budget:
          type: number
        version:
          type: integer
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateContract:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 100
          example: DeFi Bridge Integration
        description:
          type: string
          minLength: 10
          example: Integrate the Stellar bridge with our DeFi protocol
        freelancerId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        budget:
          type: number
          minimum: 0
          exclusiveMinimum: true
          example: 500
      required:
        - title
        - description
        - budget
...